% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reneuro-package.R
\docType{package}
\name{reneuro-package}
\alias{reneuro}
\alias{reneuro-package}
\title{reneuro: A European Energy System Model}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An open, extensible model of the European energy system built on 'energyRt', with the data, assumptions and processing steps kept inspectable. The current stage reproduces the core of the open PyPSA-Eur model and the workflow around it: reading networks, translating them into 'energyRt' objects, sampling in space and time, solving, and comparing the result against the source model. Planned extensions cover alternative datasets, a review of the algorithms that turn raw data into model parameters, further sectors, and links to other models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://optimal2050.github.io/reneuro/r/}
  \item \url{https://optimal2050.github.io/reneuro/}
  \item \url{https://github.com/optimal2050/reneuro}
  \item Report bugs at \url{https://github.com/optimal2050/reneuro/issues}
}

}
\author{
\strong{Maintainer}: Oleg Lugovoy \email{olugovoy@optimalsolution.dev} (\href{https://orcid.org/0000-0001-5546-9875}{ORCID})

Authors:
\itemize{
  \item Oleg Lugovoy \email{olugovoy@optimalsolution.dev} (\href{https://orcid.org/0000-0001-5546-9875}{ORCID})
}

}
\keyword{internal}
